still cant remove temporary offset in particular orientations

+++ src/com/example/stepnavi/filters/HighPassFilter.java
@@ -2,8 +2,8 @@ package com.example.stepnavi.filters;
 
 public class HighPassFilter {
 
-	private float current = 0.0f;
-	private float temp = 0.0f;
+	private double current = 0.0f;
+	private double temp = 0.0f;
 	
 	public void reset()
 	{
@@ -11,12 +11,12 @@ public class HighPassFilter {
 		temp = 	0.0f;
 	}
 	
-	public float getCurrentOutput()
+	public double getCurrentOutput()
 	{
 		return current;
 	}
 	
-	public float filter(float newValue, float magic)
+	public double filter(double newValue, double magic)
 	{
 		temp = temp*(1-magic) + magic*newValue;
 		current = newValue - temp;
